updated to fix the major vertical jitter issue caused by glitching of vsync
When vsync is inverted on LCD interface, you get a double-glitch on the vsync
due to the interrupt/reset latency at the trigger of every frame.

The FPGA is now coded to *only* accept non-inverted sync polarity from the LCD.
The sync polarity is still configurable for the HDMI stream and fully decoupled
(or should be fully decoupled) from the LCD sync polarity.