Updated FPGA to get rid of vertical jitter, finally!
Completely refactored the machine that coordinates timing between
the host LCD and the HDMI stream. A dynamic feedback loop mechanism
with smart locking is implemented. New parameters are required
to configure this, namely, a lock target and a tolerance. If you
don't set these, the image will just roll around on the screen.