1) automatically derive Km on HPD event. 2) fix setbox to not tear
Two major fixes. One is that setbox does not tear anymore, you can
call it repetetively and it will behave correctly.

The other is that now when you plug in a device, it will automatically
derive the HDCP keys. This requires a kernel change as well which
was committed separtely but may be having trouble making it into
the build system. You should see some spurious printk's now when
you plug in a device if everything is working correctly.