fixed HPD auto-HDCP mechanism; now uses semaphores. added caching.
Instead of trying to beat the race condition between the source/sink
negotiation, we cache the last Km value generated and hope that's the
one being used. In the case that it's not, we force a double-HPD event
to flush the old (wrong) Km, which causes a double-blink on new device
attach. This should mostly affect customers who swap between multiple
devices on a single display, but won't affect customers who use the
same device over and over again.

Km caching now allows devices that were previously attached to come
up seamlessly, with no screen blink.

There is one major limitation: due to HPD semaphore grabbing, there
is a 5-second window after switching to a new device that HPDs are
ignored. So if you have an N-way switchbox and you're rapidly firing through
devices it could lead to problems. This is something to test.

There is also a flaw in the semaphore implementation, the rmw lock is
not truly atomic. There is an opportunity for a race condition.
We try to eliminate it by structural mechanisms but if a user is
rapidly plugging and unplugging devices you could get unlucky and
break the semaphore lock. It's possible to make this atomic by
modifying the FPGA behavior to make the HPD bit not a simple r/w bit,
but an active bit that responds as a semaphore-request-and-lock.
This is a bit of a pain in the ass though so I'm just noting the
possibility here if we actually see this being a problem in the field.

Need to get an A/V receiver with N-way input switching to see how realistic
this scenario actually is, I don't think people "device surf" in reality
because the base HPD connect time is so awful.