Fixed bug in serial ID reporting. Fixed bug in Km caching.
Serial ID reporting had problems in the FPGA bitfile; dumb
mistakes in the verilog were made. I also
managed to tune in the timing closure on the FPGA a bit better.
The currently compiled FPGA now clocks in quite nicely.

Km caching had a bug where on disconnect/unload HPD can fire,
causing an invalid Km to be computed and cached. The invalid
value is zero, which is easy to detect, so I included a case to
catch that and abort the write in that instance. Now, Km cache
should have a useful value most of the time.