opkg: Get complete script to work wtih ro system
Due to how the complete script works -- namely that it unlinks itself
as its last step -- we need to exec a temporary script on the root
in order to remount / as ro.  This change causes the script to create
a temporary script on /, exec that script, then unlink S40complete.