major system change: matchmoded support improved
- FPGA adds signal to interrupt CPU based upon Rx PLL lock state.
This facilitates detection of the source being plugged in. The
interrupt is necessary because we want to catch the Km transaction
the moment the source is plugged in. GPIO117 (formerly the audio
in pin) was repurposed for this. Audio in isn't going to happen
anyways in this design.

- Kernel modified to report the interrupt and create two new
udev triggers (src_attach and src_detach).

- udev rules modified to signal matchmoded upon source attach,
sink attach for fast switching to overlay mode to capture the
Km event.

- matchmoded modified...a lot. It now tracks the state of
Km trigger, HPD and a few other things. Re-organized around
the new information to be smarter about the trigger. The key
challenge here is that in order for resolution changes and/or
source changes to be picked up seamlessly, matchmoded has to
be able to switch into overlay mode on an interrupt-driven
basis. Without this, the mode would linger in a 720p mode
for too long, which would cause the HDCP engine to miss the Km
event and you'll just get snow on the overly.

The 720p mode switching was also made a bit smarter based upon
further knowledge of the lock state.

Currently, there is a known issue where if an invalid resolution
is presented to NeTV, but the PLL can still lock, the device
will always pop into overlay mode. This has to do with the
"hair trigger" on overlay mode. There's a vague attempt at a routine
to force back into 720p mode, except that it upsets the PLL,
causing to lock and then unlock, which trigges us back into
overlay mode, and effectively you get a loop flopping back
and forth from overlay to 720p mode in this state. This needs
to be fixed.