fixed bug in stand-alone mode for matchmoded
A fundamental change in the FPGA behavior was required.

Originally, the FPGA would report to you, in self-timed mode,
the timing of the LCD interface going to the screen (so timings
would look sane in that mode). Unfortunately, this leads to
the bizarre problem where if the input were say, 1080i, which is
an unsupported mode but still lockable by the PLL, you would
essentially keep on jumping out of self-timed mode because
the combination of a locked PLL and valid timings from the LCD
interface makes it looks like you've got a valid 720p signal
on the input.

Now, in self-timed mode, if the input side is garbage,
the timing is garbage. This allows for clear detection of
a change back to a valid mode when in self-timed mode,
at the sacrifice of "knowing" my current resolution by
reading it out. However, you can assume safely that by
simply checking bit 3 of register 3 that if it's set,
you're in self-timed 720p mode despite what the timing
report says.

FPGA is now version 8, fpga_ctl has been updated to
correctly expect version 8.

Also fixed a bug where the PLL for self-timed mode was
being reset *before* entering self timed mode (which isn't
very useful), and created some lock-outs to deal with
concurrent hot-plugs and swaps into valid modes during
normal loop processing.

The system currently passes the following tests:

720p mode sustained
720p->1080p60->720p changes overlay->self->overlay
720p->480p->720p changes resolution and locks HDCP correctly
720p->720p (different source)->720p keeps resolution and picks up new Km correctly
720p->unplugged source changes to self
unplugged source->720p changes to overlay

it has been validated against the sony BD player and the HP BD player.

much more testing must be done, but this at least covers a core matrix
of modes to handle.