Removed setbox support from FPGA. FPGA version 9. Setbox changed.
Setbox doesn't fail to run for backward compatibility, but it
outputs a warning message. It should be possible to do a version
check in setbox as well to not affect FPGA version 9 or later but
for now it will try to do its magic on any FPGA version that exists.

FPGA version 9 checked in, fpga_ctl updated to reflect that version.
PR version bumped.

For now, we always affect the entire screen.