Major refactor of matchmoded.
Now that I better understand the problem, the core loop of
matchmoded has been rewritten.

The new paradigm is around an event-driven loop. Basically,
the asynchronous events created by kill signals from udev
drive the major state changes of the device. Thus the nexus
of power is within the handlers, and they are deputized to
do larger and more extensive things.

The function of the daemon is now just to clean up after
the events. If a change in resolution occurs, the system
notifies the browser a couple seconds after the actual hardware
change happens. if a change in lock state or invalid state
occurs, the system will more quickly change the LCD state,
but be very light-fingered about switching into a self-timed
mode. In fact, the average delay for going to self-timed mode
is on the order of 5-8 seconds.

The reason for the long delay to self-timed mode is that
many boxes go through terrible birthing pains for the connection,
flopping about between various valid and invalid modes. Some
boxes even "step" through resolutions (i.e., they will always
start in mode 1 and then graduate to the final mode as part
of the connection process). Any algorithm observing this would
just see a train wreck and decide to tell you that the device
is unsupported. Thus, the new mode of operation is to very
conservatively wait for weird hardware to recover its wits.

In fact, some hardware will remain in an infinite connection
loop under certain conditions of incompatibility. If that
happens, NeTV will forever remain in an invalid mode. But
I think that's better than thrashing to track a poorly designed
source's mode changes.

This daemon has lots of concurrency that it wasn't originally
designed for. Copious semaphores and locks are provided but
they aren't "interrupt-safe", i.e., there are zones where
you can break the read-modify-write cycle and potentially
get into a bad state. A better implementation would be to
find a way to properly create kernel-managed semaphores whose
state setting is uninterruptable. However, I don't know
how to do that so at the moment we're flapping a little bit
loose. The code is made defensively against this, so that
activities that could incur new interrupts are temporally
spaced from points where semaphores are set, so as to avoid
race conditions. However, since a source can unpredictably
change its mode, there is always a remote chance of failure.
To guard against that most semaphores will auto-reset after
a while, so in the worst case you just transition into an
odd mode for a moment, but then you come back.