make non-CEA mode detection more robust
matchmoded was triggering on random errors in timing analysis,
which would cause the image to jump periodically on the screen.

matchmoded now double-checks if the image is non-CEA timing and
the readback result for timing must match absolutely precisely
twice in a row before it will accept the timing as valid.