major change: dynamically derive EDID based on reported resolution
This introduces a subsystem where "on first boot" a function is called
that derives an EDID based upon the monitor that the NeTV is plugged into.

The program customize_edid reads the native EDID from the TV, compares
it against the resolutions supported in matchmoded_timings.h, and creates
an EDID based upon the intersection of the two.

The resulting EDID is written to /psp/cached.edid, where the netv_service
script checks if it exists and is valid, and if it is does not run the
program again.

This means every time you factory reset it'll have to regenerate the EDID.
That's a feature.

A call from the control panel to "automatically adjust settings" should
also be added. This call would simply invoke customize_edid, and everything
should work just like magic from there.