fixed some bugs in customize_edid. Modified startup script.
customize_edid now always runs on startup. It's quick and
probably the most reliable way to always have the right EDID
on the device.

this replaces the call to modeline, more or less, although
it writes the value to a cached location so you could in theory
still use modeline to set up the TV if you wanted to do it
really fast without an HPD cycle.