increase update frequency to 30 minutes, 15 minute spread
For Beta, update frequency increased to polling on 30-minute cycles
minutes, with a 15-minute spread on the poll time. Thus, updates
are guaranteed to be initiated in the worst case within
45 minutes (wait a full half hour for the cron job to run plus
15 minutes for the update to trigger); leaving about a roughly
one-hour cycle for an update to run that's not a Qt update.

This will allow a population of about 100 units to be manually
updated within a span of two work days, given 10 stations for
powering them up.

Given that the maximum population of units will be about 100
for the next month, this is a fine trade-off as there is no way
the server will fall over even if the full population was
up and running.