ok, looks like there are two problems going on...
one is that EDID derivation is traumatic. So probably a settling
time is needed after doing that; just 2 seconds allocated.

However, a loop of doing hotplug engage/disengage yields about
a 10% failure rate for correctly grabbing the cipher synchronization
state. This is worse than before. So there is some other issue
going on that's compounding the startup issue.