netv-recovery: Switch how filesystem is built
Before, we'd rely on the kernel to build a cpio image internally.
This is a bit of a problem, because our /init would run without any
/dev entries, which in turn would mean it couldn't write out debug
messages.  Aside from making this very difficult to debug, it also
meant that the restoration itself didn't work.

This patch actually builds an image under fakeroot.  This image has
/dev/ entries, and therefore works.  It is, essentially, what we
had in earlier attempts by inheriting "image", but without a lot
of the magic that came with it, and the confusing assumptions it
made when including netv-recovery in an image of its own.