linux-silvermoon-netv: Bump PR to revert LCD sign extend
Production kernels got a patch that enables sign extension.  The PR
was never bumped, so the debug builds never picked this up.  When
the release build was made, it picked up the sign-extend patch and
broke.

This will force a rebuild on production and debug, so the two will
be in sync.