networkmanager: Greatly increase scan delay
The wifi card has trouble scanning while maintaining a network
connection.  This is because it has no dedicated scanning hardware,
and must actually shift channels in order to scan.

This patch modifies NetworkManager so that it only performs a scan
once per hour when it's associated.  When disassociated, the scan
time will remain the same (usually on the order of once every
20 seconds.)

This will not affect programs such as ap_scan, because they will
continue to utilize the list of APs that NetworkManager has already
discovered.

This seems to improve packet loss, where a packet almost always
seems to get lost during a scan.