networkmanager: Only scan for APs once every hour
The hardware in the wifi card doesn't support scanning while
associated, so the kernel changes channels whenever it does a scan,
scans, then changes back to the AP it's associated with.

Occasionally this can cause missed beacons, and very strict access
points will disassociate the station if that happens.

Because we don't often move around, the list of access points can
be assumed to be relatively fixed.  Therefore, while associated,
the attached patch causes polling to occur only once an hour when
associated with an AP.